fix(psm): duplicate startEconCharter
